requirements.txt
Flask==3.0.0
groq==0.9.0
opencv-python==4.9.0.80
numpy>=1.26.3
Pillow>=10.2.0
python-dotenv==1.0.1